swapping to another auth

+++ src/main/java/org/example/melodymatch/account/AccountCredentialsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AccountCredentialsRepository extends JpaRepository<AccountCredentials, Integer> {
+public interface AccountCredentialsRepository extends JpaRepository<AccountCredentials, Long> {
     @Query("SELECT a FROM AccountModel a WHERE a.phoneNumber.phoneNumber = :phoneNumber")
     Optional<AccountCredentials>findByPhoneNumber_PhoneNumber(String phoneNumber);
 }
